Stats_Monitor/client_py/requirements.txt
psutil
bleak
pystray
Pillow
GPUtil
# pyobjc-framework-Quartz # for macOS GPU/Temp (optional)
